Fix #6 and add unit tests support
Fixing #6 was pretty easy; in fact, double quoting WITH-RSS optional
arguments was all we had to do.  Adding unit tests support to the module
on the other end, wound up being not as easy as I initially thought it
would, but still, I am quite happy with the final result.

I split the tests support in 3 different sub-activities:

1) Add a package (and a system) for the tests: `:xml-emitter/tests`
2) Add a Makefile to run the tests

  This turned out to be more annoying than expected:

  - The interactive debugger needs to be disabled when tests fail, or the
  CI job would hang forever
  - The spawn REPL needs to be shut down when tests pass, or again,
  it would cause the CI job to hang forever
  - Certain implementations did not exit with non-0 code when tests
  failed

  And all this needed to be done outside of the testing system (hence in
  the Makefile) becasue I still wanted to leave things interactive when
  people tried to run the tests with `(asdl:test-system ...)`

3) Add a .travis.yml file to test the library with multiple
   implementations

This is it!